test(*): add tempFile function that generates a truly unique filename
Previously we just used deleteme() which doesn't work that well if you run
tests in parallel or need multiple files in one testcase. As such switch to using
a sha1UUID made out of the line where the tempFile function is called and the
full path of the source file to avoid filename conflicts.