feat(corecollector): lock coredump config file when invoked for writing
This way we can ensure that we don't have to processes writing to the
file at the same time or one of them reading the file, doing its
operations but in the meantime another process also reads the file
and then overrides what the former process just wrote to the file.
This especially happened with big coredumps or when CoredumpDir.rotateDir()
took a little to clean up.